Rename rogue device to match VHDL module name and start updating for latest code
Still incomplete
